@@ -2,7 +2,7 @@
 
 	<PropertyGroup>
 		<OutputType>Exe</OutputType>
-		<TargetFramework>net5.0</TargetFramework>
+		<TargetFramework>net6.0-windows</TargetFramework>
 		<RootNamespace>Stonylang</RootNamespace>
 		<StartupObject>Stonylang.Program</StartupObject>
 		<AssemblyName>Stonylang</AssemblyName>
