@@ -1,10 +1,12 @@
-﻿using Stonylang.Lexer;
+﻿using Stonylang.Builtin;
+using Stonylang.Lexer;
 using Stonylang.Parser;
 using Stonylang.Symbols;
 using Stonylang.Utility;
 using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
+using System.Linq;
 
 namespace Stonylang.Binding
 {
@@ -12,6 +14,7 @@ namespace Stonylang.Binding
     {
         private BoundUnaryOperator(SyntaxKind kind, BoundUnaryOpKind opKind, TypeSymbol operandType)
             : this(kind, opKind, operandType, operandType) { }
+
         private BoundUnaryOperator(SyntaxKind kind, BoundUnaryOpKind opKind, TypeSymbol operandType, TypeSymbol resultType)
         {
             Kind = kind;
@@ -31,6 +34,8 @@ namespace Stonylang.Binding
             new(SyntaxKind.Not, BoundUnaryOpKind.LogicalNegation, TypeSymbol.Int),
             new(SyntaxKind.Plus, BoundUnaryOpKind.Identity, TypeSymbol.Int),
             new(SyntaxKind.Minus, BoundUnaryOpKind.Negation, TypeSymbol.Int),
+            new(SyntaxKind.Plus, BoundUnaryOpKind.Identity, TypeSymbol.Float),
+            new(SyntaxKind.Minus, BoundUnaryOpKind.Negation, TypeSymbol.Float),
             new(SyntaxKind.Inv, BoundUnaryOpKind.Inv, TypeSymbol.Int),
             new(SyntaxKind.Increment, BoundUnaryOpKind.Increment, TypeSymbol.Int),
             new(SyntaxKind.Decrement, BoundUnaryOpKind.Decrement, TypeSymbol.Int)
@@ -49,8 +54,10 @@ namespace Stonylang.Binding
     {
         private BoundBinaryOperator(SyntaxKind kind, BoundBinaryOpKind opKind, TypeSymbol type)
             : this(kind, opKind, type, type, type) { }
+
         private BoundBinaryOperator(SyntaxKind kind, BoundBinaryOpKind opKind, TypeSymbol operandType, TypeSymbol type)
             : this(kind, opKind, operandType, operandType, type) { }
+
         private BoundBinaryOperator(SyntaxKind kind, BoundBinaryOpKind opKind, TypeSymbol leftType, TypeSymbol rightType, TypeSymbol resultType)
         {
             Kind = kind;
@@ -79,7 +86,6 @@ namespace Stonylang.Binding
 
             new(SyntaxKind.Plus, BoundBinaryOpKind.Addition, TypeSymbol.Bool, TypeSymbol.String),
 
-
             // int
             new(SyntaxKind.Plus, BoundBinaryOpKind.Addition, TypeSymbol.Int),
             new(SyntaxKind.Minus, BoundBinaryOpKind.Subtraction, TypeSymbol.Int),
@@ -100,13 +106,52 @@ namespace Stonylang.Binding
             new(SyntaxKind.Less, BoundBinaryOpKind.Less, TypeSymbol.Int, TypeSymbol.Bool),
             new(SyntaxKind.LessEq, BoundBinaryOpKind.LessEq, TypeSymbol.Int, TypeSymbol.Bool),
 
+            new(SyntaxKind.Plus, BoundBinaryOpKind.Addition, TypeSymbol.Int, TypeSymbol.Float, TypeSymbol.Float),
+            new(SyntaxKind.Minus, BoundBinaryOpKind.Subtraction, TypeSymbol.Int, TypeSymbol.Float, TypeSymbol.Float),
+            new(SyntaxKind.Star, BoundBinaryOpKind.Multiplication, TypeSymbol.Int, TypeSymbol.Float, TypeSymbol.Float),
+            new(SyntaxKind.Slash, BoundBinaryOpKind.Division, TypeSymbol.Int, TypeSymbol.Float, TypeSymbol.Float),
+            new(SyntaxKind.Power, BoundBinaryOpKind.Power, TypeSymbol.Int, TypeSymbol.Float, TypeSymbol.Float),
+
+            new(SyntaxKind.Greater, BoundBinaryOpKind.Greater, TypeSymbol.Int, TypeSymbol.Float, TypeSymbol.Bool),
+            new(SyntaxKind.GreaterEq, BoundBinaryOpKind.GreaterEq, TypeSymbol.Int, TypeSymbol.Float, TypeSymbol.Bool),
+            new(SyntaxKind.Less, BoundBinaryOpKind.Less, TypeSymbol.Int, TypeSymbol.Float, TypeSymbol.Bool),
+            new(SyntaxKind.LessEq, BoundBinaryOpKind.LessEq, TypeSymbol.Int, TypeSymbol.Float, TypeSymbol.Bool),
+
             new(SyntaxKind.Plus, BoundBinaryOpKind.Addition, TypeSymbol.Int, TypeSymbol.String, TypeSymbol.String),
 
+            // float
+            
+            new(SyntaxKind.Plus, BoundBinaryOpKind.Addition, TypeSymbol.Float),
+            new(SyntaxKind.Minus, BoundBinaryOpKind.Subtraction, TypeSymbol.Float),
+            new(SyntaxKind.Star, BoundBinaryOpKind.Multiplication, TypeSymbol.Float),
+            new(SyntaxKind.Slash, BoundBinaryOpKind.Division, TypeSymbol.Float),
+            new(SyntaxKind.Power, BoundBinaryOpKind.Power, TypeSymbol.Float),
+
+            new(SyntaxKind.EqEq, BoundBinaryOpKind.LogicalEq, TypeSymbol.Float, TypeSymbol.Bool),
+            new(SyntaxKind.NotEq, BoundBinaryOpKind.LogicalNotEq, TypeSymbol.Float, TypeSymbol.Bool),
+            new(SyntaxKind.Greater, BoundBinaryOpKind.Greater, TypeSymbol.Float, TypeSymbol.Bool),
+            new(SyntaxKind.GreaterEq, BoundBinaryOpKind.GreaterEq, TypeSymbol.Float, TypeSymbol.Bool),
+            new(SyntaxKind.Less, BoundBinaryOpKind.Less, TypeSymbol.Float, TypeSymbol.Bool),
+            new(SyntaxKind.LessEq, BoundBinaryOpKind.LessEq, TypeSymbol.Float, TypeSymbol.Bool),
+
+            new(SyntaxKind.Plus, BoundBinaryOpKind.Addition, TypeSymbol.Float, TypeSymbol.Int, TypeSymbol.Float),
+            new(SyntaxKind.Minus, BoundBinaryOpKind.Subtraction, TypeSymbol.Float, TypeSymbol.Int, TypeSymbol.Float),
+            new(SyntaxKind.Star, BoundBinaryOpKind.Multiplication, TypeSymbol.Float, TypeSymbol.Int, TypeSymbol.Float),
+            new(SyntaxKind.Slash, BoundBinaryOpKind.Division, TypeSymbol.Float, TypeSymbol.Int, TypeSymbol.Float),
+            new(SyntaxKind.Power, BoundBinaryOpKind.Power, TypeSymbol.Float, TypeSymbol.Int, TypeSymbol.Float),
+
+            new(SyntaxKind.Greater, BoundBinaryOpKind.Greater, TypeSymbol.Float, TypeSymbol.Int, TypeSymbol.Bool),
+            new(SyntaxKind.GreaterEq, BoundBinaryOpKind.GreaterEq, TypeSymbol.Float, TypeSymbol.Int, TypeSymbol.Bool),
+            new(SyntaxKind.Less, BoundBinaryOpKind.Less, TypeSymbol.Float, TypeSymbol.Int, TypeSymbol.Bool),
+            new(SyntaxKind.LessEq, BoundBinaryOpKind.LessEq, TypeSymbol.Float, TypeSymbol.Int, TypeSymbol.Bool),
+
+            new(SyntaxKind.Plus, BoundBinaryOpKind.Addition, TypeSymbol.Float, TypeSymbol.String, TypeSymbol.String),
 
             // string
             new(SyntaxKind.Plus, BoundBinaryOpKind.Addition, TypeSymbol.String),
             new(SyntaxKind.Plus, BoundBinaryOpKind.Addition, TypeSymbol.String, TypeSymbol.Bool, TypeSymbol.String),
             new(SyntaxKind.Plus, BoundBinaryOpKind.Addition, TypeSymbol.String, TypeSymbol.Int, TypeSymbol.String),
+            new(SyntaxKind.Plus, BoundBinaryOpKind.Addition, TypeSymbol.String, TypeSymbol.Float, TypeSymbol.String),
             new(SyntaxKind.Star, BoundBinaryOpKind.Multiplication, TypeSymbol.String, TypeSymbol.Int, TypeSymbol.String),
             new(SyntaxKind.Minus, BoundBinaryOpKind.Subtraction, TypeSymbol.String, TypeSymbol.Int, TypeSymbol.String),
 
@@ -129,6 +174,7 @@ namespace Stonylang.Binding
         private readonly SourceText _source;
         private BoundScope _scope;
         public DiagnosticBag Diagnostics => _diagnostics;
+
         public Binder(SourceText source, BoundScope parentScope)
         {
             _scope = new(parentScope);
@@ -138,6 +184,7 @@ namespace Stonylang.Binding
         public static BoundGlobalScope BindGlobalScope(BoundGlobalScope previous, SourceText source, CompilationUnitSyntax syntax)
         {
             BoundScope parentScope = CreateParentScope(previous);
+
             Binder binder = new(source, parentScope);
             BoundStmt stmt = binder.BindStatement(syntax.Statement);
             ImmutableArray<VariableSymbol> variables = binder._scope.GetDeclaredVariables();
@@ -156,17 +203,25 @@ namespace Stonylang.Binding
                 previous = previous.Previous;
             }
 
-            BoundScope parent = null;
+            BoundScope parent = CreateRootScope();
             while (stack.Count > 0)
             {
                 previous = stack.Pop();
                 BoundScope scope = new(parent);
-                foreach (var v in previous.Variables) scope.TryDeclare(v, out var _);
+                foreach (var v in previous.Variables) scope.TryDeclareVariable(v);
 
                 parent = scope;
             }
 
             return parent;
+        }
+
+        private static BoundScope CreateRootScope()
+        {
+            BoundScope result = new(null);
+            foreach (FunctionSymbol f in BuiltinFunctions.GetAll())
+                result.TryDeclareFunction(f);
+            return result;
         }
 
         public BoundStmt BindStatement(StmtNode stmt) => stmt.Kind switch
@@ -195,7 +250,14 @@ namespace Stonylang.Binding
         private BoundVariableStmt BindVariableStmt(VariableStmt stmt)
         {
             BoundExpr initializer = BindExpression(stmt.Initializer);
-            return new(BindVariable(stmt.Identifier, initializer.Type, stmt.IsMut), initializer);
+            TypeSymbol type = stmt.Type == null ? initializer.Type : LookUpType(stmt.Type.Lexeme);
+            if (type == null)
+            {
+                _diagnostics.Report(_source, stmt.Type.Span, $"Unknown type \"{stmt.Type.Lexeme}\".", "TypeException", LogLevel.Error);
+                type = TypeSymbol.Error;
+            }
+            initializer = BindConversion(stmt.Initializer, type, false);
+            return new(BindVariable(stmt.Identifier, type, stmt.IsMut), initializer);
         }
 
         private BoundIfStmt BindIfStmt(IfStmt stmt)
@@ -207,7 +269,7 @@ namespace Stonylang.Binding
         private BoundWhileStmt BindWhileStmt(WhileStmt stmt)
         {
             BoundExpr condition = stmt.Condition == null ? new BoundLiteralExpr(true) : BindExpression(stmt.Condition, TypeSymbol.Bool);
-            return new(condition, BindBlockStmt(stmt.ThenBranch), stmt.IsDoWhile);
+            return new(condition, BindBlockStmt(stmt.ThenBranch), stmt.DoKeyword != null);
         }
 
         private BoundForStmt BindForStmt(ForStmt stmt)
@@ -223,29 +285,34 @@ namespace Stonylang.Binding
             return new(variable, initialValue, range, body);
         }
 
-        private BoundExpressionStmt BindExpressionStmt(ExpressionStmt stmt) => new(BindExpression(stmt.Expression));
+        private BoundExpressionStmt BindExpressionStmt(ExpressionStmt stmt) => new(BindExpression(stmt.Expression, true));
 
-        public BoundExpr BindExpression(ExprNode expr, TypeSymbol expectedType)
+        public BoundExpr BindExpression(ExprNode expr, bool canBeVoid = false)
         {
-            BoundExpr expression = BindExpression(expr);
-            if (expression.Type != TypeSymbol.Error &&
-                expectedType != TypeSymbol.Error &&
-                expression.Type != expectedType)
-                _diagnostics.Report(_source, expr.Span, $"Cannot convert type of {expression.Type} to type of {expectedType}.", "TypeException", LogLevel.Error);
-            return expression;
+            BoundExpr boundExpr = BindExpressionInternal(expr);
+            if (!canBeVoid && boundExpr.Type == TypeSymbol.Void)
+            {
+                _diagnostics.Report(_source, expr.Span, $"Expression must have a non-void value.", "TypeException", LogLevel.Error);
+                return new BoundErrorExpr();
+            }
+
+            return boundExpr;
         }
 
-        public BoundExpr BindExpression(ExprNode expr) => expr.Kind switch
+        public BoundExpr BindExpression(ExprNode expr, TypeSymbol expectedType) => BindConversion(expr, expectedType);
+        public BoundExpr BindExpressionInternal(ExprNode expr) => expr.Kind switch
         {
             SyntaxKind.LiteralExpr => BindLiteralExpr((LiteralExpr)expr),
             SyntaxKind.UnaryExpr => BindUnaryExpr((UnaryExpr)expr),
             SyntaxKind.BinaryExpr => BindBinaryExpr((BinaryExpr)expr),
             SyntaxKind.NameExpr => BindNameExpr((NameExpr)expr),
             SyntaxKind.AssignmentExpr => BindAssignmentExpr((AssignmentExpr)expr),
+            SyntaxKind.CallExpr => BindCallExpr((CallExpr)expr),
             _ => throw new Exception($"Unexpected syntax \"{expr.Kind}\"."),
         };
 
         private static BoundExpr BindLiteralExpr(LiteralExpr expr) => new BoundLiteralExpr(expr.Value ?? 0);
+
         private BoundExpr BindUnaryExpr(UnaryExpr expr)
         {
             BoundExpr boundOperand = BindExpression(expr.Operand);
@@ -259,7 +326,7 @@ namespace Stonylang.Binding
                 {
                     if (boundOperand is BoundVariableExpr bv)
                     {
-                        if (_scope.TryLookUp(bv.Variable.Name, out var variable))
+                        if (_scope.TryLookUpVariable(bv.Variable.Name, out var variable))
                         {
                             if (!variable.IsMut)
                                 _diagnostics.Report(_source, expr.Op.Span, $"Cannot assign to \"{bv.Variable.Name}\" since it is a read-only variable.", "AssignmentExcption", LogLevel.Error);
@@ -299,7 +366,7 @@ namespace Stonylang.Binding
             if (expr.Name.IsMissing)
                 return new BoundErrorExpr();
 
-            if (_scope.TryLookUp(expr.Name.Lexeme, out var value))
+            if (_scope.TryLookUpVariable(expr.Name.Lexeme, out var value))
                 return new BoundVariableExpr(value);
 
             if (expr.Name.Kind != SyntaxKind.EOF)
@@ -312,7 +379,7 @@ namespace Stonylang.Binding
             string name = expr.Name.Lexeme;
             BoundExpr boundExpresion = BindExpression(expr.Value);
 
-            if (!_scope.TryLookUp(name, out var variable))
+            if (!_scope.TryLookUpVariable(name, out var variable))
             {
                 if (expr.Name.Kind != SyntaxKind.EOF)
                     _diagnostics.Report(_source, expr.Name.Span, $"Could not find \"{name}\" in the current context.", "KeyNotFoundException", LogLevel.Error);
@@ -322,24 +389,87 @@ namespace Stonylang.Binding
             if (!variable.IsMut)
                 _diagnostics.Report(_source, expr.Name.Span, $"Cannot assign to \"{name}\" since it is a read-only variable.", "AssignmentException", LogLevel.Error);
 
-            if (!variable.Type.Equals(boundExpresion.Type))
+            BoundExpr boundConversion = BindConversion(expr.Value, variable.Type, false);
+
+            return new BoundAssignmentExpr(new(name, boundExpresion.Type, null, expr.EqualsToken.Span), boundConversion);
+        }
+
+        private BoundExpr BindCallExpr(CallExpr expr)
+        {
+            if (expr.Arguments.Count == 1 && LookUpType(expr.Identifier.Lexeme) is TypeSymbol type)
+                return BindConversion(expr.Arguments[0], type, true);
+
+            IEnumerable<FunctionSymbol> functions = BuiltinFunctions.GetAll();
+            if (!_scope.TryLookUpFunction(expr.Identifier.Lexeme, out var function))
             {
-                _diagnostics.Report(_source, expr.Name.Span, $"Cannot assign a value with type of \"{boundExpresion.Type}\" to \"{name}\", which has a type of \"{variable.Type}\".", "TypeException", LogLevel.Error);
-                return boundExpresion;
+                _diagnostics.Report(_source, expr.Identifier.Span, $"Function \"{expr.Identifier.Lexeme}\" doesn't exist in the current context.", "KeyNotFoundException", LogLevel.Error);
+                return new BoundErrorExpr();
             }
 
-            return new BoundAssignmentExpr(new(name, boundExpresion.Type, null, expr.EqualsToken.Span), boundExpresion);
+            ImmutableArray<BoundExpr> boundArguments = (from i in expr.Arguments
+                                                        select BindExpression(i)).ToImmutableArray();
+
+            if (boundArguments.Length != function.Parameters.Length)
+            {
+                for (int i = 0; i < function.Parameters.Length; i++)
+                {
+                    if (boundArguments.ElementAtOrDefault(i) == null || function.Parameters[i].Type != boundArguments[i].Type)
+                        _diagnostics.Report(_source, expr.Arguments.LastOrDefault<Node>(expr.LParen).Span,
+                            $"There is no given argument that corresponds to the required formal parameter \"{function.Parameters[i].Name}\" of \"{function.Name}\".", "ArgumentException", LogLevel.Error);
+                }
+
+                if (boundArguments.Length > function.Parameters.Length)
+                    _diagnostics.Report(_source, expr.Arguments.LastOrDefault<Node>(expr.LParen).Span,
+                        $"Expected {function.Parameters.Length} arguments, but got {boundArguments.Length}.", "ArgumentException", LogLevel.Error);
+
+                return new BoundErrorExpr();
+            }
+
+            return new BoundCallExpr(function, boundArguments);
+        }
+
+        private BoundExpr BindConversion(ExprNode expr, TypeSymbol type, bool allowExplicit = false)
+        {
+            BoundExpr boundExpr = BindExpression(expr);
+            Conversion conversion = Conversion.Classify(boundExpr.Type, type);
+            if (!conversion.Exists)
+            {
+                if (boundExpr.Type != TypeSymbol.Error &&
+                    type != TypeSymbol.Error &&
+                    boundExpr.Type != type)
+                    _diagnostics.Report(_source, expr.Span,
+                        $"Cannot convert a value of type \"{boundExpr.Type}\" to a value of type \"{type}\".", "TypeException", LogLevel.Error);
+                return new BoundErrorExpr();
+            }
+
+            if (!allowExplicit && conversion.IsExplicit)
+                _diagnostics.Report(_source, expr.Span,
+                    $"Cannot implicitly convert a value of type \"{boundExpr.Type}\" to a value of type \"{type}\". An explicit conversion exists; are you missing a cast?", "TypeException", LogLevel.Error);
+
+            /*if (conversion.IsIdentity)
+            {
+                if (conversionInfo)
+                    _diagnostics.Report(_source, expr.Span, $"Unnecessary conversion from type \"{type}\" to \"{type}\"", "TypeInfo", LogLevel.Info);
+                return boundExpr;
+            }*/
+            return new BoundConversionExpr(type, boundExpr);
         }
 
         private VariableSymbol BindVariable(Token identifier, TypeSymbol type, bool isMut)
         {
-            string name = identifier.Lexeme ?? "?";
-            bool declare = !identifier.IsMissing;
-            VariableSymbol variable = new(name, type, null, identifier.Span, isMut);
-
-            if (declare && !_scope.TryDeclare(variable, out _))
-                _diagnostics.Report(_source, identifier.Span, $"Variable \"{identifier.Lexeme}\" was already declared in the current or a previous scope.", "DeclarationException", LogLevel.Error);
+            VariableSymbol variable = new(identifier.Lexeme ?? "?", type, null, identifier.Span, isMut);
+            if (!identifier.IsMissing && !_scope.TryDeclareVariable(variable) || _scope.TryLookUpFunction(identifier.Lexeme, out _))
+                _diagnostics.Report(_source, identifier.Span, $"\"{identifier.Lexeme}\" was already declared in the current or a previous scope.", "DeclarationException", LogLevel.Error);
             return variable;
         }
+
+        private static TypeSymbol LookUpType(string name) => name switch
+        {
+            "bool" => TypeSymbol.Bool,
+            "int" => TypeSymbol.Int,
+            "float" => TypeSymbol.Float,
+            "string" => TypeSymbol.String,
+            _ => null,
+        };
     }
 }
