@@ -1,35 +1,43 @@
 ﻿using Stonylang.Utility;
-using System;
+using System.Collections.Immutable;
 
 namespace Stonylang.Symbols
 {
     public enum SymbolKind
     {
         Type,
-        Variable
+        Variable,
+        Function,
+        Parameter
     }
 
     public abstract class Symbol
     {
-        private protected Symbol(string name) => Name = name;
+        protected Symbol(string name) => Name = name;
+
         public string Name { get; }
         public abstract SymbolKind Kind { get; }
+
         public override string ToString() => Name;
     }
 
     public sealed class TypeSymbol : Symbol
     {
         public static readonly TypeSymbol Error = new("unknown");
+        public static readonly TypeSymbol Any = new("any");
         public static readonly TypeSymbol Bool = new("bool");
         public static readonly TypeSymbol Int = new("int");
+        public static readonly TypeSymbol Float = new("float");
         public static readonly TypeSymbol String = new("string");
+        public static readonly TypeSymbol Void = new("void");
+
         internal TypeSymbol(string name)
             : base(name) { }
 
         public override SymbolKind Kind => SymbolKind.Type;
     }
 
-    public sealed class VariableSymbol : Symbol
+    public class VariableSymbol : Symbol
     {
         public VariableSymbol(string name, TypeSymbol type, object value, TextSpan? span, bool isMut = false)
             : base(name)
@@ -40,10 +48,36 @@ namespace Stonylang.Symbols
             IsMut = isMut;
         }
 
+        public override SymbolKind Kind => SymbolKind.Variable;
         public TypeSymbol Type { get; }
         public object Value { get; set; }
         public TextSpan? Span { get; }
         public bool IsMut { get; }
-        public override SymbolKind Kind => SymbolKind.Variable;
+    }
+
+    public sealed class ParameterSymbol : VariableSymbol
+    {
+        public ParameterSymbol(string name, TypeSymbol type, TextSpan? span, bool isMut = false)
+            : base(name, type, null, span, isMut)
+        {
+        }
+
+        public override SymbolKind Kind => SymbolKind.Parameter;
+    }
+
+    public sealed class FunctionSymbol : Symbol
+    {
+        public FunctionSymbol(string name, ImmutableArray<ParameterSymbol> parameters, TypeSymbol type, TextSpan? span)
+            : base(name)
+        {
+            Parameters = parameters;
+            Type = type;
+            Span = span;
+        }
+
+        public override SymbolKind Kind => SymbolKind.Function;
+        public ImmutableArray<ParameterSymbol> Parameters { get; }
+        public TypeSymbol Type { get; }
+        public TextSpan? Span { get; }
     }
 }
