@@ -1,12 +1,17 @@
 ﻿using Stonylang.Lexer;
+using Stonylang.Utility;
 
 namespace Stonylang.Parser
 {
-    public abstract class ExprNode : Node { }
+    public abstract class ExprNode : Node
+    { }
 
     public sealed class LiteralExpr : ExprNode
     {
-        public LiteralExpr(Token literalToken) : this(literalToken, literalToken.Literal) { }
+        public LiteralExpr(Token literalToken) : this(literalToken, literalToken.Literal)
+        {
+        }
+
         public LiteralExpr(Token literalToken, object value)
         {
             LiteralToken = literalToken;
@@ -49,6 +54,7 @@ namespace Stonylang.Parser
     public sealed class NameExpr : ExprNode
     {
         public NameExpr(Token name) => Name = name;
+
         public override SyntaxKind Kind => SyntaxKind.NameExpr;
         public Token Name { get; }
     }
@@ -66,5 +72,22 @@ namespace Stonylang.Parser
         public Token Name { get; }
         public Token EqualsToken { get; }
         public ExprNode Value { get; }
+    }
+
+    public sealed class CallExpr : ExprNode
+    {
+        public CallExpr(Token identifier, Token lParen, SeparatedSyntaxList<ExprNode> arguments, Token rParen)
+        {
+            Identifier = identifier;
+            LParen = lParen;
+            Arguments = arguments;
+            RParen = rParen;
+        }
+
+        public override SyntaxKind Kind => SyntaxKind.CallExpr;
+        public Token Identifier { get; }
+        public Token LParen { get; }
+        public SeparatedSyntaxList<ExprNode> Arguments { get; }
+        public Token RParen { get; }
     }
 }
